chore: detach rnet client from abstui
