+++ src/Director/LingoEngine.Director.Core/Remote/Commands/ConnectRNetClientCommand.cs
@@ -1,0 +1,7 @@
+using AbstUI.Commands;
+
+namespace LingoEngine.Director.Core.Remote.Commands;
+
+/// <summary>Connects the Director RNet client to the host.</summary>
+public sealed record ConnectRNetClientCommand() : IAbstCommand;
+
