+++ src/Director/LingoEngine.Director.Core/Remote/Commands/DisconnectRNetClientCommand.cs
@@ -1,0 +1,7 @@
+using AbstUI.Commands;
+
+namespace LingoEngine.Director.Core.Remote.Commands;
+
+/// <summary>Disconnects the Director RNet client from the host.</summary>
+public sealed record DisconnectRNetClientCommand() : IAbstCommand;
+
