+++ src/Director/LingoEngine.Director.Core/UI/DirectorMainMenu.cs
@@ -22,7 +22,6 @@ using LingoEngine.Director.Core.Remote.Commands;
 using LingoEngine.Net.RNetHost;
 using LingoEngine.Net.RNetHost.Commands;
 using LingoEngine.Net.RNetClient;
-using LingoEngine.Net.RNetClient.Commands;
 using System;
 using System.ComponentModel;
