--- src/Net/LingoEngine.Net.RNetClient/Commands/ConnectRNetClientCommand.cs
@@ -1,7 +1,0 @@
-using AbstUI.Commands;
-
-namespace LingoEngine.Net.RNetClient.Commands;
-
-/// <summary>Connects the RNet client to the host.</summary>
-public sealed record ConnectRNetClientCommand() : IAbstCommand;
-
