--- src/Net/LingoEngine.Net.RNetClient/Commands/DisconnectRNetClientCommand.cs
@@ -1,7 +1,0 @@
-using AbstUI.Commands;
-
-namespace LingoEngine.Net.RNetClient.Commands;
-
-/// <summary>Disconnects the RNet client from the host.</summary>
-public sealed record DisconnectRNetClientCommand() : IAbstCommand;
-
